task 0

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,7 +7,7 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t number = 0;
+	unsigned int number = 0;
 
 	if (!h)
 		return (0);
